+++ lib/generators/boring/factory_bot/install/templates/users.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :user do
+    email { "boring+generators@example.com" }
+    password { "boring+generators" }
+  end
+end