+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## master (unreleased)
 * Adds Flipper generator. ([@abhaynikam][])
+* Adds FactoryBot generator. ([@abhaynikam][])
+* Update default node-version for GitHub Action generator and fixes PG setup issues. ([@abhaynikam][])
 
 ## 0.11.0 (June 23rd, 2021)
 * Adds Stimulus generator. ([@abhaynikam][])
 * Adds Rails Admin generator. ([@abhaynikam][])
 * Adds Paper Trail generator. ([@abhaynikam][])
-* Update default node-version for GitHub Action generator and fixes PG setup issues. ([@abhaynikam][])
 
 ## 0.10.0 (May 26th, 2021)
 * Updates Stripe payment generator. ([@abhaynikam][])